xdma_thread: fix cpu node bug
The number of threads is arbitrarily set to 8 and we iterate over 8 cpus
without knowing if we really have 8 cpus. This causes a call to
cpu_to_node with an index that is out of bound, thus returning an
undefined number, which crashes later when calling
kthread_create_on_node.
Fix this by iterating over the online cpus stopping when reaching the
number of threads that is specified.